Numpy segmentation forced to be uint8. Int16 is not supported by updateSegmentBinaryLabelmapFromArray function

+++ CBCTToothSegmentation/CBCTToothSegmentation.py
@@ -505,7 +505,7 @@ class CBCTToothSegmentationLogic(ScriptedLoadableModuleLogic):
         # process prediction output
         output = inferer(inputProcessed, model)
         output = torch.softmax(output, axis=1).data.cpu().numpy()
-        output = np.argmax(output, 1).squeeze().astype(int)
+        output = np.argmax(output, 1).squeeze().astype(np.uint8)
 
         # Crop the predicion back to original size
         lower = [0]*3
